Move requiredDeviceSize to the main Payload class (#1297905)
dnf is not the only payload type.
