+++ pyanaconda/packaging/dnfpayload.py
@@ -597,20 +597,6 @@ class DNFPayload(packaging.PackagePayload):
         log.debug("Total size required %s", total_space)
         return total_space
 
-    def requiredDeviceSize(self, format_class):
-        """ We need to provide information how big device is required to have successful
-            installation. ``format_class`` should be filesystem format
-            class for the **root** filesystem this class carry information about
-            metadata size.
-
-            :param format_class: Class of the filesystem format.
-            :type format_class: Class which inherits :class:`blivet.formats.fs.FS`
-            :returns: Size of the device with given filesystem format.
-            :rtype: :class:`blivet.size.Size`
-        """
-        device_size = format_class.getRequiredSize(self.spaceRequired)
-        return device_size.roundToNearest(Size("1 MiB"))
-
     def _isGroupVisible(self, grpid):
         grp = self._base.comps.group_by_pattern(grpid)
         if grp is None:
